+++ src/template/User/UserTitle.js
@@ -1,0 +1,14 @@
+import { getter } from '../../store/index.js';
+
+const UserTitle = () => {
+  const userName = getter.userName();
+  const getTitleName = userName ? `${ userName }\'s` : '';
+
+  return `
+    <h1 id="user-title" data-username="${userName}">
+      <span><strong>${ getTitleName }</strong>Todo List</span>
+    </h1>
+  `;
+};
+
+export default UserTitle;
